Face Detection
Why are we using OpenCV? I think in the simplest terms is that OpenCV is the most prominent and widely used computer vision related library. In addition with it’s support of Python and C++ along with it’s features it allows for algorithms such as face detection to be implemented efficiently.